% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{standardise}
\alias{standardise}
\title{Standardises a vector: more details to come!}
\usage{
standardise(x, ...)
}
\arguments{
\item{x}{The vector}

\item{...}{Other arguments}
}
\value{
The standardised vector
}
\description{
Standardises a vector: more details to come!
}
\examples{
standardise(1:20)
}
